Fixes from linting

+++ tests/test_generators.py
@@ -965,20 +965,24 @@ def test_extruding_hypnotoad_mesh_to_wall_remesh() -> None:
         restrict_to_vessel=True,
         mesh_to_core=True,
         mesh_to_wall=True,
-        wall_resolution=1.,
+        wall_resolution=1.0,
     )
 
     assert len(mesh.reference_layer.bounds) == 2
     assert len(mesh.reference_layer.bounds[0]) == 0
 
     def distance(start: SliceCoord, end: SliceCoord) -> float:
-        return np.sqrt((start.x1 - end.x1) ** 2 + (start.x2 - end.x2)**2)
+        return float(np.sqrt((start.x1 - end.x1) ** 2 + (start.x2 - end.x2) ** 2))
 
     wall_segment_lengths = [
-        distance(*quad.shape([0., 1.0]).iter_points()) for quad in mesh.reference_layer.bounds[1]
+        distance(*quad.shape([0.0, 1.0]).iter_points())
+        for quad in mesh.reference_layer.bounds[1]
     ]
     initial_wall = hypno_mesh.equilibrium.wall
-    initial_wall_length = min(np.sqrt((p1.R - p0.R)**2 + (p1.Z - p0.Z)**2) for p0, p1 in itertools.pairwise(initial_wall))
+    initial_wall_length = min(
+        np.sqrt((p1.R - p0.R) ** 2 + (p1.Z - p0.Z) ** 2)
+        for p0, p1 in itertools.pairwise(initial_wall)
+    )
     assert max(wall_segment_lengths) / min(wall_segment_lengths) <= 3
     assert all(length < initial_wall_length for length in wall_segment_lengths)
